requirements.txt

Flask
google-cloud-speech
pydub
ffmpeg
